requirements.txt
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
-f https://download.pytorch.org/whl/torch_stable.html
pandas==1.1.1
numpy==1.18.1
Flask==1.1.1
nltk==3.4.5
networkx==2.3
afinn==0.1
beautifulsoup4==4.8.2
newspaper3k==0.2.8
validators==0.18.2
torch==1.6.0+cpu
torchvision==0.7.0+cpu
sentence-transformers==0.3.9
gunicorn
